+++ src/where/index.ts
@@ -7,14 +7,15 @@ export interface WhereQuery<_Model> {
   value: any
 }
 
-export type RegularWhereFilter = Exclude<
+export type BasicWhereFilter = Exclude<
   FirestoreWhereFilterOp,
-  'array-contains'
+  'array-contains' | 'in'
 >
 
+// Basic filter variation
 function where<Model, Key extends keyof Model>(
   field: Key | [Key],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key]
 ): WhereQuery<Model>
 // array-contains variation
@@ -28,10 +29,17 @@ function where<
   filter: 'array-contains',
   value: ValueArray[ValueType]
 ): WhereQuery<Model>
+// in variation
+function where<Model, Key extends keyof Model>(
+  field: Key | [Key],
+  filter: 'in',
+  value: Model[Key][]
+): WhereQuery<Model>
 
+// Basic filter variation
 function where<Model, Key1 extends keyof Model, Key2 extends keyof Model[Key1]>(
   field: [Key1, Key2],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key1][Key2]
 ): WhereQuery<Model>
 // array-contains variation
@@ -46,7 +54,14 @@ function where<
   filter: 'array-contains',
   value: ValueArray[ValueType]
 ): WhereQuery<Model>
+// in variation
+function where<Model, Key1 extends keyof Model, Key2 extends keyof Model[Key1]>(
+  field: [Key1, Key2],
+  filter: 'in',
+  value: Model[Key1][Key2][]
+): WhereQuery<Model>
 
+// Basic filter variation
 function where<
   Model,
   Key1 extends keyof Model,
@@ -54,7 +69,7 @@ function where<
   Key3 extends keyof Model[Key1][Key2]
 >(
   field: [Key1, Key2, Key3],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key1][Key2][Key3]
 ): WhereQuery<Model>
 // array-contains variation
@@ -70,7 +85,19 @@ function where<
   filter: 'array-contains',
   value: ValueArray[ValueType]
 ): WhereQuery<Model>
+// in variation
+function where<
+  Model,
+  Key1 extends keyof Model,
+  Key2 extends keyof Model[Key1],
+  Key3 extends keyof Model[Key1][Key2]
+>(
+  field: [Key1, Key2, Key3],
+  filter: 'in',
+  value: Model[Key1][Key2][Key3][]
+): WhereQuery<Model>
 
+// Basic filter variation
 function where<
   Model,
   Key1 extends keyof Model,
@@ -79,7 +106,7 @@ function where<
   Key4 extends keyof Model[Key1][Key2][Key3]
 >(
   field: [Key1, Key2, Key3, Key4],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key1][Key2][Key3][Key4]
 ): WhereQuery<Model>
 // array-contains variation
@@ -96,7 +123,20 @@ function where<
   filter: 'array-contains',
   value: ValueArray[ValueType]
 ): WhereQuery<Model>
+// in variation
+function where<
+  Model,
+  Key1 extends keyof Model,
+  Key2 extends keyof Model[Key1],
+  Key3 extends keyof Model[Key1][Key2],
+  Key4 extends keyof Model[Key1][Key2][Key3]
+>(
+  field: [Key1, Key2, Key3, Key4],
+  filter: 'in',
+  value: Model[Key1][Key2][Key3][Key4][]
+): WhereQuery<Model>
 
+// Basic filter variation
 function where<
   Model,
   Key1 extends keyof Model,
@@ -106,7 +146,7 @@ function where<
   Key5 extends keyof Model[Key1][Key2][Key3][Key4]
 >(
   field: [Key1, Key2, Key3, Key4, Key5],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key1][Key2][Key3][Key4][Key5]
 ): WhereQuery<Model>
 // array-contains variation
@@ -124,7 +164,21 @@ function where<
   filter: 'array-contains',
   value: ValueArray[ValueType]
 ): WhereQuery<Model>
+// in variation
+function where<
+  Model,
+  Key1 extends keyof Model,
+  Key2 extends keyof Model[Key1],
+  Key3 extends keyof Model[Key1][Key2],
+  Key4 extends keyof Model[Key1][Key2][Key3],
+  Key5 extends keyof Model[Key1][Key2][Key3][Key4]
+>(
+  field: [Key1, Key2, Key3, Key4, Key5],
+  filter: 'in',
+  value: Model[Key1][Key2][Key3][Key4][Key5][]
+): WhereQuery<Model>
 
+// Basic filter variation
 function where<
   Model,
   Key1 extends keyof Model,
@@ -135,7 +189,7 @@ function where<
   Key6 extends keyof Model[Key1][Key2][Key3][Key4][Key5]
 >(
   field: [Key1, Key2, Key3, Key4, Key5, Key6],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key1][Key2][Key3][Key4][Key5][Key6]
 ): WhereQuery<Model>
 // array-contains variation
@@ -154,7 +208,22 @@ function where<
   filter: 'array-contains',
   value: ValueArray[ValueType]
 ): WhereQuery<Model>
+// in variation
+function where<
+  Model,
+  Key1 extends keyof Model,
+  Key2 extends keyof Model[Key1],
+  Key3 extends keyof Model[Key1][Key2],
+  Key4 extends keyof Model[Key1][Key2][Key3],
+  Key5 extends keyof Model[Key1][Key2][Key3][Key4],
+  Key6 extends keyof Model[Key1][Key2][Key3][Key4][Key5]
+>(
+  field: [Key1, Key2, Key3, Key4, Key5, Key6],
+  filter: 'in',
+  value: Model[Key1][Key2][Key3][Key4][Key5][Key6][]
+): WhereQuery<Model>
 
+// Basic filter variation
 function where<
   Model,
   Key1 extends keyof Model,
@@ -166,7 +235,7 @@ function where<
   Key7 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6]
 >(
   field: [Key1, Key2, Key3, Key4, Key5, Key6, Key7],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7]
 ): WhereQuery<Model>
 // array-contains variation
@@ -186,7 +255,23 @@ function where<
   filter: 'array-contains',
   value: ValueArray[ValueType]
 ): WhereQuery<Model>
+// in variation
+function where<
+  Model,
+  Key1 extends keyof Model,
+  Key2 extends keyof Model[Key1],
+  Key3 extends keyof Model[Key1][Key2],
+  Key4 extends keyof Model[Key1][Key2][Key3],
+  Key5 extends keyof Model[Key1][Key2][Key3][Key4],
+  Key6 extends keyof Model[Key1][Key2][Key3][Key4][Key5],
+  Key7 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6]
+>(
+  field: [Key1, Key2, Key3, Key4, Key5, Key6, Key7],
+  filter: 'in',
+  value: Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][]
+): WhereQuery<Model>
 
+// Basic filter variation
 function where<
   Model,
   Key1 extends keyof Model,
@@ -199,7 +284,7 @@ function where<
   Key8 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7]
 >(
   field: [Key1, Key2, Key3, Key4, Key5, Key6, Key7, Key8],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8]
 ): WhereQuery<Model>
 // array-contains variation
@@ -220,7 +305,24 @@ function where<
   filter: 'array-contains',
   value: ValueArray[ValueType]
 ): WhereQuery<Model>
+// in variation
+function where<
+  Model,
+  Key1 extends keyof Model,
+  Key2 extends keyof Model[Key1],
+  Key3 extends keyof Model[Key1][Key2],
+  Key4 extends keyof Model[Key1][Key2][Key3],
+  Key5 extends keyof Model[Key1][Key2][Key3][Key4],
+  Key6 extends keyof Model[Key1][Key2][Key3][Key4][Key5],
+  Key7 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6],
+  Key8 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7]
+>(
+  field: [Key1, Key2, Key3, Key4, Key5, Key6, Key7, Key8],
+  filter: 'in',
+  value: Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8][]
+): WhereQuery<Model>
 
+// Basic filter variation
 function where<
   Model,
   Key1 extends keyof Model,
@@ -234,7 +336,7 @@ function where<
   Key9 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8]
 >(
   field: [Key1, Key2, Key3, Key4, Key5, Key6, Key7, Key8, Key9],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8][Key9]
 ): WhereQuery<Model>
 // array-contains variation
@@ -256,7 +358,25 @@ function where<
   filter: 'array-contains',
   value: ValueArray[ValueType]
 ): WhereQuery<Model>
+// in variation
+function where<
+  Model,
+  Key1 extends keyof Model,
+  Key2 extends keyof Model[Key1],
+  Key3 extends keyof Model[Key1][Key2],
+  Key4 extends keyof Model[Key1][Key2][Key3],
+  Key5 extends keyof Model[Key1][Key2][Key3][Key4],
+  Key6 extends keyof Model[Key1][Key2][Key3][Key4][Key5],
+  Key7 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6],
+  Key8 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7],
+  Key9 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8]
+>(
+  field: [Key1, Key2, Key3, Key4, Key5, Key6, Key7, Key8, Key9],
+  filter: 'in',
+  value: Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8][Key9][]
+): WhereQuery<Model>
 
+// Basic filter variation
 function where<
   Model,
   Key1 extends keyof Model,
@@ -271,7 +391,7 @@ function where<
   Key10 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8][Key9]
 >(
   field: [Key1, Key2, Key3, Key4, Key5, Key6, Key7, Key8, Key9, Key10],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8][Key9][Key10]
 ): WhereQuery<Model>
 // array-contains variation
@@ -291,8 +411,26 @@ function where<
   ValueType extends keyof ValueArray
 >(
   field: [Key1, Key2, Key3, Key4, Key5, Key6, Key7, Key8, Key9, Key10],
-  filter: RegularWhereFilter,
+  filter: BasicWhereFilter,
   value: ValueArray[ValueType]
+): WhereQuery<Model>
+// in variation
+function where<
+  Model,
+  Key1 extends keyof Model,
+  Key2 extends keyof Model[Key1],
+  Key3 extends keyof Model[Key1][Key2],
+  Key4 extends keyof Model[Key1][Key2][Key3],
+  Key5 extends keyof Model[Key1][Key2][Key3][Key4],
+  Key6 extends keyof Model[Key1][Key2][Key3][Key4][Key5],
+  Key7 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6],
+  Key8 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7],
+  Key9 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8],
+  Key10 extends keyof Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8][Key9]
+>(
+  field: [Key1, Key2, Key3, Key4, Key5, Key6, Key7, Key8, Key9, Key10],
+  filter: 'in',
+  value: Model[Key1][Key2][Key3][Key4][Key5][Key6][Key7][Key8][Key9][Key10][]
 ): WhereQuery<Model>
 
 /**
@@ -316,7 +454,7 @@ function where<
  * ```
  *
  * @param field - The field or key path to query
- * @param filter - The filter operation ('<', '<=', '==', '>=' or '>')
+ * @param filter - The filter operation ('<', '<=', '==', '>=', '>', 'array-contains', 'in' or 'array-contains-any')
  * @param value - The value to pass to the operation
  * @returns The where query object
  */